# backend/server/swagger.yaml
openapi: 3.0.3
info:
  title: Pixel Protocol API
  description: |
    Backend API for Pixel Protocol game with hero management and battles.
    
    ## Authentication
    Most endpoints require player authentication via:
    1. **Request Body**: Include `_secret` field with player's UUID secret (primary method)
    2. **Cookie**: Use `player_secret` cookie set by hero creation (fallback method)
    
    When creating a hero, a secure `player_secret` cookie is automatically set for subsequent requests.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.pixelprotocol.com
    description: Production server
paths:
...
  /api/hero:
    post:
      summary: Create a new hero
      description: |
        Creates a new hero and sets a secure `player_secret` cookie for future authentication.
        This cookie can be used instead of providing `_secret` in request bodies for subsequent requests.
      tags:
        - Hero
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - _secret
              properties:
                title:
                  type: string
                  example: "Dragon Slayer"
                description:
                  type: string
                  example: "A mighty warrior who defeats dragons"
                _secret:
                  type: string
                  format: uuid
                  description: Player's secret for authentication
      responses:
        '201':
          description: Hero created successfully and authentication cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "player_secret=uuid-value; Path=/; HttpOnly; Secure; SameSite=Strict"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '400':
          description: Invalid request
        '401':
          description: Invalid secret
        '500':
          description: Internal server error
  /api/hero/{id}:
    get:
      summary: Get hero by ID
      tags:
        - Hero
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretRequest'
      responses:
        '200':
          description: Hero details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '404':
          description: Hero not found
          
    put:
      summary: Update hero (full update)
      tags:
        - Hero
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - _secret
              properties:
                title:
                  type: string
                description:
                  type: string
                _secret:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Hero updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '404':
          description: Hero not found
        '401':
          description: Unauthorized
          
    patch:
      summary: Update hero (partial update)
      tags:
        - Hero
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - _secret
              properties:
                title:
                  type: string
                description:
                  type: string
                _secret:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Hero updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '404':
          description: Hero not found
        '401':
          description: Unauthorized
          
    delete:
      summary: Soft delete hero
      tags:
        - Hero
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretRequest'
      responses:
        '204':
          description: Hero deleted
        '404':
          description: Hero not found
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    PlayerSecret:
      type: apiKey
      in: header
      name: Cookie
      description: |
        Player authentication via secure cookie. Cookie is automatically set when creating a hero.
        Cookie name: `player_secret`
    PlayerSecretBody:
      type: http
      scheme: bearer
      description: |
        Player authentication via request body. Include `_secret` field with player's UUID secret.
        Takes precedence over cookie authentication.
  schemas:
    Player:
      type: object
      properties:
        ID:
          type: string
          format: uuid
        UserName:
          type: string
        UserNameSuffix:
          type: integer
          format: int32
        Secret:
          type: string
          format: uuid
        DeletedAt:
          type: string
          format: date-time
          nullable: true
    Hero:
      type: object
      properties:
        ID:
          type: string
          format: uuid
        Country:
          type: string
          example: "US"
        Elo:
          type: integer
          format: int32
          example: 1000
        Title:
          type: string
        Description:
          type: string
        PlayerID:
          type: string
          format: uuid
        Player:
          $ref: '#/components/schemas/Player'
        DeletedAt:
          type: string
          format: date-time
          nullable: true
...
    SecretRequest:
      type: object
      required:
        - _secret
      properties:
        _secret:
          type: string
          format: uuid
          description: Player's secret for authentication
...